Update to current

+++ Assignment1/src/ObjectAllocator.cpp
@@ -58,7 +58,7 @@ using word_t = intptr_t ;
 /******************************************************************************/
 constexpr size_t operator "" _z(unsigned long long n)
 {
-    return n;
+    return static_cast<size_t>(n);
 }
 
 /******************************************************************************/
